Create branch o1-trust-score-queries. The purpose of this branch is to speed up queries of the type: getTrust(Identity truster, Identity trustee) getScore(Identity truster, Identity trustee)
Because db4o does not support multi-dimensional indices, those queries
currently take O(N), where N is the number of trust-/score-objects in
the database.
Those queries CAN be executed in O(1) with the modifcations of this
branch.
Since we use these queries heavily in Score-computation (and therefore
trustlist import), the old score-computation code would run in O(N^2
* ...), where the new code possibly runs in O(N * ...).