Initial implementation of caching for being able to resolve the following database queries quickly: Getting trust between two identities; getting score between two identities; getting an identity by ID.
NOTICE:
This branch is currently broken, do not use it
The unit test fail because score computation fails. Most likely this
happens because the cache is not flushed if a Score/Trust object is
deleted. (The identity cache would also have to be flushed when an
identity is deleted but the side effects of that are not visible in unit
test yet).