FCPInterface.ClientTrackerDaemon: Remove garbage collection thread
- The fred code of branch plugin-fcp-rewrite will provide this as an
API, no need to implement it in WOT anymore. (See
class FCPPluginClientTracker there)
- The get() and put() are kept since I'm not sure whether the fred code
will be usable in unit tests yet. So if we change ClientTrackerDaemon to
be an interface to FCPPluginClientTracker, it will be possible to make
it use the HashMaps in unit tests in case the fred code does not work
there.