FCPInterface.ClientTrackerDaemon: Finish basic implementation of using plugin-fcp-rewrite
- Implement get() to use the new fred API:
* Return FCPPluginClient as thats the new representation of the remote
end instead of PluginReplySender. Use UUID as thats the fred-side
ID-type for FCPPluginClient
* Throw IOException since thats what the fred getter throws. Notice that
IOException is what shall be typically thrown in Java if a network
connection was closed meanwhile, so this is the right thing here.
- Leave put() empty as fred stores FCPPluginClient objects
automatically. But keep it for now since the fred stuff probably won't
work in unit tests yet. Document  the unit test issue with a FIXME.
- Add FIXME for removing the class as it merely is a wrapper around the
fred stuff now. Again, keep it for the unit tests.
- Add FIXME about the obsolete JavaDoc