FCPInterface: Begin implementation of new interface FredPluginFCPMessageHandler.ServerSideFCPMessageHandler
First steps here are to:
- Rename the message handler function
- Make sure it does not process reply messages because those should not
arrive here: WOT only does send messages to the client which shall
receive a reply using *synchronous* sends. Synchronous sends will have
their reply returned to the blocking send function instead of having it
passed to the message handler. Anything else sent by WOT should not
receive a reply. Thus, reply messages should not arrive in the
async message handling function.

The rest of the body of the function will be adapted in the following
commits.