FCPInterface: Specify deprecation mechanism for fields of the SimpleFieldSets which WOT sends
- The plugin-fcp-rewrite of fred will have special standardized
errorCode and errorMessage values natively at fred plugin FCP messages.
Thus, the following commits which adapt the
FCPInterface.errorMessageFCP() will have to deprecate the old error code
and message fields which were previously put  into the SimpleFieldSet of
the message in a non-standardized way. So it is a good time to
standardize a deprecation mechanism for SFS entries of WOT.
- I chose to include the deprecated field list in the actual on-network
data to ensure that client authors read them :) Also, it will make it
easier to understand dumps of raw FCP traffic. For example during the
RFC of event-notifications, I had supplied a dump to the mailing list.
But as it contained lot of deprecated fields, people got confused and
had to ask about the duplicate fields and which ones are the right ones
etc.