FCPInterface.send*ChangedNotification(): Adapt to new throws declaration of sendChangeNotification()
- Also add final attribute to variables so all of the 3 functions have
the same signature.
- FCPInterface.java should compile again now, so the basic adaption to
fred plugin-fcp-rewrite is finished for this file. A full review
of it should be done nevertheless to ensure absolute compatibility with
plugin-fcp-rewrite. But I'll first compile fix everything and just run
the unit tests to see whether it works. If it does, the review can maybe
be spared. Notice that the FIXMEs in the file need for sure be resolved
before the branch can be merged though.