FCPInterface: Throw out a specific Exception type if sendSynchronous() succeed but processing at the remote side did not
The previous code did use IOException for both:
- sendSynchronous() failing because the connectin is dead
- sendSynchronous() succeeding, returning a reply, but the reply havign
success==false which says that the client had an error at its side.

I had previously assumed that this distinction would be of no use to the
SubscriptionManager which uses the affected functions. I thought it
would only matter to it whether it has to re-send a certain message. But
looking at the actual code now shows that it in fact already can do two
things: Terminate the Subscription immediately if the connection is
closed, and not terminate but re-send an event if only processing
failed.
So we need to use two Exception types.