SubscriptionManager: Throw out InterruptedException during synchronization of a Subscription
- This fixes storeNewSubscriptionAndCommit() and its callers from
converting the InterruptedException to RuntimeException.
- Synchronization can take a very long time as we are possibly sending
out the whole WOT database to the client. To prevent the client from
hammering WOT during shutdown with re-subscribing immediately, we need
to pass out the InterruptedException to it so it can honor it by
terminating as well.
- The commit also improves the Exception handling at
storeNewSubscriptionAndCommit() to deal with specific exceptions instead
of having a catch-all block
- This breaks compilation of the FCPInterface as it doesn't handle that
yet. Will add handling there in the following commits.