SubscriptionManager.Client.sendNotifications(): Catch more specific Exceptions
- Specify which kinds of Exception can happen instead of catching all at
once.
- Even though the specification of the function which is called in the
try{} does not allow RuntimeException it is a good idea to guard against
RuntimeException here to separate bugs of the FCP interface from having
the SubscriptionManager bug out. The sendNotifications() function here
is a better place to do that than the called notifySubscriberByFCP()
since there exist 3 implementations of notifySubscriberByFCP() already,
so there would be a lot of code duplication. Also, the
sendNotifications() is what is capable of queueing the stuff for
re-sending upon error, so it should deal with errors.