SubscriptionManager.Client.sendNotifications(): Improve logging
- FCPCallFailedException is thrown by the FCP interface if we received
an actual reply message from the client which confirms that our message
did arrive but the client had an error processing it at its side. So
this is not an error at our side and thus should not be logged as such.
We log it as WARNING instead of NORMAL because it still is a problem
which should be debugged.
- RuntimeException is not allowed by the specification of
sendNotifications() and thus always is an error and must be logged as
such.