SubscriptionManager: Add performance TODO
Currently, all notifications are deployed one by one in a single
thread which loops over all clients. This causes each client to block
deployment for all of the other clients. A better concept would be to
have one thread per client.
Notice that this TODO is also enforced by the bugtracker entry
https://bugs.freenetproject.org/view.php?id=6350 which requests the
developers to search the source code for "TODO: Optimization" soon
enough.