FCPInterface.handleSubscribe(): Catch InterruptedException
This was recently added to the throws declaration of
SubscriptionManager.subscribeToIdentities() / subscribeToTrusts() /
subscribeToScores().

We have to honor it by NOT returning a reply to the client:
- If we returned success, that would be a lie, and we shouldn't lie to
the client about being properly connected to WOT. This could cause false
display in the UI, and false bug reports because people would be
wondering why their client receives no data even though it is connected.
- If we did return failure, the client could immediately retry filing
the Subscription, which is a heavy operation, and thus block shutdown.
So we must return no reply to ensure that the client code which waits
for a success/failure reply times out to delay the retry attempt
implicitely.

Notice that this also adapts the caller of handeSubscribe() to NOT
create an success reply automatically because handleSubscribe()
returned null.