SubscriptionManagerFCPTest: Begin adapting to plugin-fcp-rewrite
- Interface PluginReplySender was part of the old pre-rewrite API. It
was used to send replies from the server plugin to the client. In the
case of this test, it was implemented to store the server messages so
the test can process them. We now change it to implement the new
interface for client-side FCP message handler, because it was being used
like a client.
- Because the interface it now implements has a more fitting name, i.e.
a name which indicates that it is a client, this commit removes the
JavaDoc which explained in detail that it is a reply receiver and not a
reply sender.
- The handlePluginFCPMessage() could now store the FCPPluginMessage
instead of only the SimpleFieldSet params because that contains more
information than the SimpleFieldSet, namely a success indicator and an
errorCode. So we add a FIXME to store it and use it in tests. First
steps is the tests to compile and succeed as is though, so this is only
a FIXME.
- The JavaDoc of handlePluginFCPMessage() instructs it to always return
a reply message if the original message was not a reply, so we do that.
I've reviewed the whole class, it doesn't seem to do anything which
needs it to indicate failure to the server, so we always return a
success reply.