SubscriptionManagerFCPTest.fcpCall(): Adapt to fred plugin-fcp-rewrite
- As I want to use the stuff for an implicit test of fred's
plugin-fcp-rewrite, this is  changed to use the actual FCPPluginClient
of fred to send the message instead of calling the message handler
directly.
- Because we used to call the message handler directly which would have
only returned after the message handler is finished, we now have to
use the blocking sendSynchronous() instead of non-blocking send() to
ensure that the reply message has  been delivered before fcpCall()
returns. Or in other words: We cannot use send() because it would queue
the sending in another thread, so the reply might not be available yet
when it returns.
- WOT should compile again now :) Next step of adapting to fred
plugin-fcp-rewrite is to run the unit tests and debug eventual issues.
After that, a DebugFCPClient test run should be done.