Unit tests: Add draft of class for unit tests to run inside a full node
This will be needed to fix SubscriptionManagerFCPTest to work with fred
plugin-fcp-rewrite. The issues which makes us require a node are:
- The SubscriptionManager needs to be able to query FCP clients by ID so
it can send events to them even for a long time after the client has
opened the connection. It cannot store the clients in the database
because they are fred objects, and because it needs to NOT keep a strong
references to them so they can be GCed in case they disconnect.
- Fred plugin-fcp-rewrite provides the by-ID storage, to replace WOT's
own by-ID storage.
- But to access fred's by-ID client storage, we need to access the fred
PluginRespirator, which in turn accesses the fred FCPPluginClienTracker,
which is created by the FCPServer. So there are quite a lot of fred
classes interacting, and I think its easier to just create a fest node
than to try to emulate the creation of all of them.
- It will also be a benefit for future unit tests to be able to use a
full node I suppose.