Ant builder: Put unit tests into JAR
- I want to add unit tests which run a full Freenet node. This requires
WOT to be loaded as a plugin, so we have to compile a JAR.
- The location to which tests are compiled is changed from
"${build-test}" to "${build-test}/test" so we can stick the JAR into
${build-test} without it being on its own include path. I decide against
putting it into dist/ because the distribution should only contain stuff
which *has* been tested. Also there is no sense in distributing unit
test JARs IMHO.
- The <jar> tag is inspired by the one in the "dist" <target>.
- The <zipfileset> which includes the unit test classes in the test
runner includes "**/*Test*.class" instead of "**/*.class" because
including all files would pass the main WOT classes to JUnit, which
would fail because they are not tests.
So we use the heuristic of havin anything be a test which has "Test" in
the name. I have grepped the directory listing of all test classes, and
they currently all have "Test" in the name, so this will succeed for
the existing classes. If a developer adds a new test, he is very likely
to try to run it and check for its results - so its likely that
people will notice if new tests do not get included because they don't
have "Test" in the name.
- A quick Google search resulted in the claim that <jvmarg> is safe to
include unqouted ${} paths as long as we only use one in one <jvmarg>
tag. I have not tested this with a path which contains spaces, but will
do once a test is implemented which actually uses the JAR.