AbstractFullNodeTest: Implement basic node creation & WOT loading
- Create a node with as much disabled as possible. This might break
stuff, but thats not very relevant to WOT as long as the WOT tests
succeed, which we will see when running them after they have been
adapted.
- Load the WOT plugin from the filame which is passed to the JVM as
specified in build.xml (via the property WOT_test_jar)
- Add basic sanity check whether the WOT database is empty to prevent
it from persisting across test restarts. This does not guarantee that
the node's files are properly deleted during test shutdown so a FIXME
to check that is added