AbstractFullNodeTest: Fix test failure due to IndexOutOfBoundsException
- This was because the web interface of fred is not available in a unit
test node: WOT would try to register its web interface there, which
obviously won't work if fred has no web interface.
- The fix is to disable the WOT web interface if Freenet's web interface
is disabled.
- Notice that the get("fproxy") needs no null check because the whole
"fproxy.enabled" config option is registered in the SimpleToadletServer
during startup even if it is disabled, so it will be set to its defaulf
if it is not in the freenet.ini file. This is also why I think it *is*
clean code to use the string names of config options instead of getters:
If one made a typo and used the wrong name, it would throw a
NullPointerException. So the registration provides the same level of
code safety with Strings as using normal Java code would. Also notice
that there seems to be no getter yet, I did check that, so we have to
use  the String getters for now anyway.