AbstractFullNodeTest: Fix self test failure
- It would fail because the WOT database *will* contain Identitys even
at first run: WOT will automatically create seed Identitys
- It was a bad test anyway: If the database persisted, it would only
fail at the *second* test run. This is a design failure, tests should be
self-contained. Instead, now we check whether the node user data
directory is deleted at tearDown(). To do so, we terminate the node and
then check whether the directory exists. Notice that this will fail
currently as the node doesn't seem to delete it, I am committing it
before deletion is implemented so the test results are correct already.
Also, for double safety we check whether the directory exists during
setUp() - to guard against crashed test runs which did not delete the
directory.