AbstractFullNode: Implement auto deletion of node directory
- Uses JUnit4's class TemporaryFolder. Nice that it has this! The class
will automatically delete the folder. Also it will fail if it exists
already, I have manually validated that for JUnit 4.11 which seems to be
the most recent and also in my package manager. Therefore, the asserts
which validate that it is deleted and doesn't exist can be removed since
it is JUnit's job to work properly now.
- Also adds TODO to the old unit tests which create a database file to
use this.