SubscriptionManagerFCPTest: Start using AbstractFullNodeTest as base class
- Strip down setUp() so it only initializes the FCP connection and
everything else is done at the member variables already, so the stuff is
less spaghetti.
- As a consequence, rename setUp() to setUpClient(). JUnit4 allows
arbitrary names because what is run before the test is determined by
"@Before", not by the function name.
- Use PluginRespirator.connecToOtherPlugin() instead of
FCPPluginClient.constructForUnitTest(): connectToOtherPlugin() is the
main API for obtaining FCP connections in real nodes, and we can use it
now since a PluginRespirator is available now due to the fact that the
parent class now creates a full Freenet node. It is always better to use
production code than special code paths for unit tests!
- The mFCPInterface variable was not used anymore, so it can be removed.
- Make the class final so unused variables are detected by the compiler
the next time.
- There are unrelated compiling failures remaining, the following
commits will fix them.