Abstract*Test: Make the classes really abstract
They were not abstract since JUnit will then complain because it cannot
instantiate them, so they would have to be ignored, but that could not
be done because they contained self-tests.
But I think self-tests are also possible using member classes. So this
commit ...:
- makes them abstract
- adds an @Ignore("") which contains an explanation of how to add self
tests
- removes the cheap self-test at AbstractJUnit4BaseTest as child tests
will explode anyway if it is null, not worth converting a that trivial
one to a member class.
- does not affect any self-test at AbstractFullNodeTest because there is
no self-test in that class anymore already