AbstractJUnit4BaseTest: Copy addRandomIdentities() from deprecated DatabaseBasedTest
- Deprecates the one in DatabaseBasedTest, so copypasting it is fine as
the other one will then be removed someday due to the deprecation.
- Will be needed for fixing SubscriptionManagerFCPTest to compile
against AbstractJUnit4BaseTest being the new top level base class
instead of DatabaseBasedTest
- Adds an abstract getWebOfTrust() to AbstractJunit4BaseTest since that
class does not generate a WebOfTrust object, it is at the responsibility
of the child classes to create one. This is because child classes will
have different ways of creating one: AbstractFullNodeTest will obtain
the WebOfTrust object by loading the actual JAR into a real node; in
opposite to that the future replacement of DatabaseBasedTest will create
dummy WebOfTrust objects without a node. So it must be up to the child
class how to create a WebOfTrustObject.