AbstractJUnit4BaseTest: Compile fix
- This is the necessary adapting due to the fact that the function was
copy-psated from DatabaseBasedTest in the previous commit.
- The repository should compile again now :) Remember the goal of the
previous commits was adapting SubscriptionManagerFCPTest to compile
again as its base class was changed from DatabaseBasedTest to
AbstractFullNodeTest (and thus AbstractJUnit4BaseTest as it is parent of
AbstractFullNodeTest) as part of adapting it to work with fred branch
plugin-fcp-rewrite. Next step will be to debug the test results of
SubscriptionManagerFCPTest  now running against plugin-fcp-rewrite.
(Albeit I will probably do some commits in between to add line-wrapping
and JavaDoc to the new stuff in AbstractJUnit4BaseTest).