AbstractFullNodeTest: Ignore return value of NodeStarter.globalTestInit()
- The function will return the same RNG as which was passed to it.
- This has been documented at its JavaDoc since
0335ae8131b16020f01dd7ffc0d9f704cea52656
- It is a good idea to ignore the return value: The function will
throw in unit tests right now because it is called more than once for
the same JVM. So we have to change the code to only call it once for the
same JVM. With the current state of saving the return value, we would
have to change the mRandom member variable to be static, which sucks.
Since the return value will be the same as was passed to it anyway, we
can igore it and should do so.