SubscriptionManagerFCPTest: Fix ComparisonFailure expected:<[0]> but was:<[5]>
The class was previously a child of DatabaseBasedTest, but now is a
child of AbstractFullNodeTest. In opposite to the previous parent class,
the new one causes WOT to create the seed identities, so the database is
not empty. The fix is to use the deleteSeedIdentities() of the parent
class to get rid of them.

The original stack trace was:
org.junit.ComparisonFailure: expected:<[0]> but was:<[5]>
	at org.junit.Assert.assertEquals(Assert.java:115)
	at org.junit.Assert.assertEquals(Assert.java:144)
	at plugins.WebOfTrust.SubscriptionManagerFCPTest.
		testSubscribeTo(SubscriptionManagerFCPTest.java:174)
	at plugins.WebOfTrust.SubscriptionManagerFCPTest.
		testSubscribeUnsubscribe(SubscriptionManagerFCPTest.java:148)
	at plugins.WebOfTrust.SubscriptionManagerFCPTest.
		testSubscribeUnsubscribe(SubscriptionManagerFCPTest.java:140)
	at sun.reflect.NativeMethodAccessorImpl.invoke0(Native Method)