FCPClientReferenceImplementation: Use FCPPluginClient instead of PluginTalker
- This is the begin of adapting the whole class to use the branch
fred-plugin-fcp-rewrite (whose primary client class is FCPPluginClient)
instead of the legacy PluginTalker stuff. Compilation is broken by this,
but it is a good idea to split up the work into multiple commits. The
last commit of the series of refactoring this class will state that it
compiles again.
- The mConnectionIdentifier is removed for the following reaons: The
connection identifier was previously used to determine whether
a reply message actually was a reply to a message which we had sent
as part of the existing FCPPluginClient mConnection, or whether it was
a reply from a previous connection which has timed out and was replaced
with a new FCPPluginClient.
This is not needed  anymore since in fcp-plugin-rewrite, the message
handling function will be passed the FCPPluginClient to which the
message was sent, so we can check whether it belongs to our
current FCPPluginClient, or an old one just by comparing to the
FCPPluginClient mConnection.