FCPClientReferenceImplementation.connect(): Adapt to plugin-fcp-rewrite
- mConnectionIdentifier was removed, so we don't need to deal with it
anymore. See previous commit.
- PluginRespirator.connecToOtherPlugin() is the designated replacement
for the deprecated getPluginTalker(). It expects to be passed an
implementor of FredPluginFCPMessageHandler.ClientSideFCPMessageHandler
instead of the legacy FredPluginTalker, so FCPMessageReceiver is changed
to implement the new API. It is NOT adapted to actually contain the new
required functions yet, that will be done in the following commits. It
won't compile like it is now, so its safe to postpone.