FCPClientReferenceImplementation.send(): Adapt to plugin-fcp-rewrite
- mConnection previously was a PluginTalker, but
now is a FCPPluginClient. So send() has a different signature
- The old function was non-blocking, so we can use the non-blocking
send() here as well
- In opposite to the old implementation, we need to decide whether we
send a reply or a non-reply message, by chosing the appropriate
FCPluginMessage constructor. I've checked all callers, and the
operations they do are not replies. So we are safe to send a non-reply
message here. However, it must be documented that this function must
only be used for non-replies, so JavaDoc is added.
- The new send() can throw IOException, so we force callers to deal with
it by throwing it out. Will adapt them in the following commits.