FCPClientReferenceImplementation.fcp_Ping(): Adapt to plugin-fcp-rewrite
- Throw out IOException which is newly thrown by send() if it detects
that the connection was closed before spawning the actual send thread.
- Handle IOException in the only caller KeepaliveLoop.run(): Log that
the connection was closed, don't do any further sending, use
force_disconnect() to mark it as closed. The already existing call to
scheduleKeepaliveLoopExecution() will then notice that mConnection is
null, and schedule the keepalive loop to be re-executed after the short
WOT_RECONNECT_DELAY. Thus, we don't need to reconnect in the current
iteration.
- Use explicit "return;" statement to make it clear that we are done.
Move the logging at the end of the function into the finally{} block to
still work then.