FCPClientReferenceImplementation.KeepaliveLoop: Make sure to always prevent thread death
The try { ...} catch(almost everything) is to prevent thread death to
ensure that the connection is always kept alive by the thread and thus
the try{} should contain all of the function. I had left some of it out
because I assumed that that code wouldn't throw but that is an arbitrary
decision.