FCPClientReferenceImplementation.checkSubscriptions(): Adapt to new throws declaration of fcp_Subscribe() / fcp_Unsubscribe()
- Throw out the IOException to the caller since we cannot do anything
about it here.
- Notice that we especially do NOT have to update the tables which
contain the "is subscribed?" state: That state is updated by two
possible things: 1) The FCP message handler, which will receive
Unsubscribed messages from WOT. That will work even if we do nothing
here. 2) force_disconnect(), which the caller of this function typically
will end up calling on its own because it makes sense: IOException means
the fred side of the connection has told us that the connection is
closed, so force_disconnect() should be used to cleanup.
- No commit to adapt its callers will follow since the only caller
already catches IOException and then does force_disconnect().