FCPClientReferenceImplementation...handlePluginFCPMessage(): Adapt to new connection state detection in plugin-fcp-rewrite
The name of the FCP plugin we talk to isn't passed directly anymore,
it is wrapped in the passed FCPPluginClient. And it doesn't really
matter what it is as we now have a clear way of checking whether we're
talking to the right client: To check whether the passed one is the one
we have stored in mConnection.
Thus, also adapt the checking of whether the message belongs to our
connection to only validate whether client == mConnection.
Notice: The previous "identifier" was being abused in the old FCP API as
a per-connection identifier even though it was meant as a per-message
identifier. So even though "identifier" still exists as part of the
FCPPluginMessage, we do not check it here.