FCPClientReferenceImplementation...handlePluginFCPMessage(): Adapt to new error signaling in plugin-fcp-rewrite
- The old pre plugin-fcp-rewrite code was using the feature of
branch PluginTalker-sync-send which allowed it to throw out a
RuntimeException  to cause that RuntimeException to be passed to the
sender of the  original message. This mechanism has been replaced by
returning a FCPPluginMessage object which was constructed as a reply to
contain success==false, which we can do by using
FCPPluginMessage.constructErrorReply().
- Also, we must only return such a reply if the original message was a
reply already - replies to replies are not allowed by JavaDoc of
FredPluginFCPMessageHandler.
- While I was it, the documetnation about the "dangling
subscription" was amended to stress that this means the *remote* code
at WOT: We're sending a reply to the remote side so it can terminate
the subscription, so the statement that it is remote should be added to
the documentation