FCPClientReferenceImplementation...handlePluginFCPMessage(): Adapt to new error signaling in plugin-fcp-rewrite
- The old pre plugin-fcp-rewrite code was using the feature of
branch PluginTalker-sync-send which allowed it to throw out a
RuntimeException to cause that RuntimeException to be passed to the
sender of the  original message. This mechanism has been replaced by
returning a FCPPluginMessage object which was constructed as a reply
to contain success==false, which we can do by using
FCPPluginMessage.constructErrorReply().
- Also, we must only return such a reply if the original message was
NOT areply already - replies to replies are not allowed by JavaDoc of
FredPluginFCPMessageHandler.
- Notice that this commit also adds returning an error reply to the
branch which handles unknown message types even though it did not throw
previous: The JavaDoc of the parent interface requests us to always
return a reply if we are allowed to, so we do now.
- Notice that we do Logger.error(this,. ..) instead of
Logger.error(FCPPluginClient.this, ...) as the rest of the function also
does that. I do think that the Logger configuration of fred will be
smart enough to detect child classes as well if people only enable
logging for FCPClientReferenceImplementation. Nevertheless, a FIXME is
added to ensure that people check this.
- With this commit, the adaption of FCPClientReferenceImplementation to
plugin-fcp-rewrite is finished as far as it can be said that it will
compile now :) Whats next is:
* doing a test run with DebugFCPClient to see whether
FCPClientReferenceImplementation still works correctly. A FIXME is added
to enforce this.
* to do a full review of the class to fix eventual JavaDoc issues and
to validate its general correctness. A FIXME is added to enforce this.