FCPInterface: Remove class ClientTrackerDaemon
- Its original purpose before fred plugin-fcp-rewrite was to store FCP
clients by ID. plugin-fcp-rewrite now provides this stored, so we don't
have to do it ourselves
- The class was not removed during adaption to plugin-fcp-rewrite
because I wasn't sure whether it would be possible to get the client
storage of fred work in unit tests. It does work in unit tests now
(thanks to AbstractFullNodeTest) so we don't need ClientTrackerDaemon
anymore
- The template for eventually joining sendSynchronous() threads during
shutdown is kept and put into FCPInterface.stop() where it belongs. It
also is adapted to be pseudo-code which matches what we should do when
waiting for sendSynchronous() threads. Before, it was shutdown code of
ClientTrackerDaemon, so it didn't quite match.