DebugFCPClient: Fix logging of possibly huge HashMaps
- When I wrote the original code, It was originally planned to log the
class of the subscription synchronization (which can be Identity /
Trust / Score) as this commit does, not the HashMap. Logging the HashMap
is dangerous because if WOT sends us a secondary synchronization, it
contains possibly the WHOLE database of WOT.
- I tried to do this with Java reflection instead, but gave up after
half an hour, its too complex given that we only use it for one place
right now.