WebOfTrust.beginTrustListImport(): Fix comitting half-complete transactions
- This happened in unit tests on branch use-plugin-fcp-rewrite. The
symptom would that createOwnIdentity() would succeed but the returned
OwnIdentity would not be stored in the database because
beginTrustListImport() rolled back the part which stores it but did not
throw so createOwnIdentity() could finish. Then the following
setTrustWithoutCommit() calls of the unit tests would fail because they
would be passed an OwnIdentity object which is not stored in the
database: Trust.storeWithoutCommit() will throw if the truster is not
stored.
- The fix is to change beginTrustListImport() to throw an Exception
when it has rolled back the transaction so the caller does not believe
that the transaction is still alive.
- I have considered writing database upgrade code to fix eventual
breakage left behind by incomplete createOwnIdentity() transactions but
all callers of beginTrustListImport() outside of the unit tests seem to
be safe, so I won't do this.