SubscriptionManagerFCPTest: Test values of new fields introduced in fred plugin-fcp-rewrite
- Namely FCPluginMessage.success, .errorCode and .errorMessage.
- testSubscribeTo(String type) does not check the errorCode yet because
the errorCode is not even set by the FCPInterface yet. The FIXME in the
FCPInterface thus is updated to tell whoever resolves it to update this
unit test as well.
- subscribeAndSynchronize() and importSynchronization() do not check the
new fields in depth because testSubscribeTo(String type) already tests
the same FCP message types. However testSubscribeTo(String type) does
NOT test the notification message types, so importNotifications() has to
test the new fields as well.