FCPClientReferenceImplementation.pingTimedOut(): Fix wrong documentation
The onReply() function doesn't even exist anymore, it was replaced by
FCPMessageReceiver.handlePluginFCPMessage() as part of adapting to
fred-plugin-fcp-rewrite.
And that function does not handle the ping reply itself, it delegates
that to the FCPPongHandler.
You can also confirm this by using Eclipse to search for all references
to that variable.