FCPInterface.handleSubscribe(): Store the error type in FCPPluginMessage.errorCode
- This replaces the old mechanism of putting the error in the
"Description" field in params.
- errorMessage is not added because as said in the comments, the
Subscribe FCP function is not typically used by UI. (errorMessage shall
typically be localized, so it would be a bit of work to add the
translation strings).
- We don't break any existing client applications with removing the
"Description" field because the whole subscriptions code has not been
released in an official WOT build yet. So clients should not be using it
yet.
- Also fixes the existing parsers of WOT, namely the
FCPClientReferenceImplementation and the unit test
SubscriptionManagerFCPTest to use the new place for the error code.