FCPInterface.handlePluginFCPMessage(): Fully document error messages
- Fix documentation of SubscriptionExistsAlready message to mention that
the identifier will be the same as in the original Subscribe message.
- Document which errorCode and other stuff is used if the user indicated
failure of processing the synchronization message. I've validated this
with a test run. This also fixes the fact that the documentation about
that error message was not int he "Errors" section, and that it looked
like the error message would be of type "Message" = "Subscribed", which
it is not.