FCPInterface.handlePluginFCPMessage(): Remove non-critical FIXME
There is already a TODO at that place to do this.
The FIXME was only added because I treat FIXMEs with higher priority
than TODOs and wanted to ensure that I do it before merging the current
branch use-plugin-fcp-rewrite. I thought this would be a good idea to do
now because what the branch does is using the new fred branch
plugin-fcp-rewrite which adds the errorCode field. But I've already
adapted the event-notifications (= handleSubscribe()) code to use the
new errorCode field, and thats our only immediate goal to get done. The
other legacy FCP functions can stay as is for now, as people shouldn't
use most of them anymore anyway one event-notifications is out.