SubscriptionManager: Begin splitting up synchronizeSubscriberByFCP()
The motivation behind this is this FIXME:
https://github.com/freenet/plugin-WebOfTrust/blob/use-plugin-fcp-rewrite/src/plugins/WebOfTrust/ui/fcp/FCPInterface.java#L80
Summary: Currently, synchronizeSubscriberByFCP() does two things:
1) Read a full copy of all Identity/Trust/Score objects in the WOT
database.
2) Send it by FCP to the subscribing Client.

This obviously can take a very long time: We are potentially reading the
whole database, AND then sending it to the client by network. Especially
are we sending it with the sendSynchronous() function of the
FCPPluginClient - so the send call will block until all data has been
transfered.

To speed this up, the commit adds a second, abstract function to the
Subscription class: storeSynchronization(). It will do the above step
(1) = cloning the database. The old function
synchronizeSubscriberByFCP()'s JavaDoc is changed to require it to only
do the job of sending the clone by FCP.

NOTICE: This does NOT yet provide an implementation, fix the
implementation of synchronizeSubscriberByFCP() or the callers of it.
Thus, this starts the branch split-up-subscription-synchronization,
which will deal with necessary remaining adaptions.