SubscriptionManager: Add skeleton of storage class for synchronizations.
- This provides storage for the mechanism whose implementation began in
the previous commit 339b4646e3866239a6f1e06572a6507dc487f946.
- The class is introduced to ensure that the Notification class does not
have to be modified: We can now pass objects of this class to
Notification pretending they are the "newObject" of the Notification.
This is nice because modifying Notification itself to be able to store a
synchronization would have sucked: It pretty much relies on having an
oldObject and newObject, adding the synchronization as something
different would complexify its fine checks of validity of oldObject and
newObject.