SubscriptionManager.storeNewSubscriptionAndCommit(): Implement usage of storeSynchronizationWithoutCommit().
- The synchronization has to be stored at the time of creation of the
Subscription: It has to happen before and right before the first
Notification is stored to ensure that the client is completely in sync.
So if we store it at the creation of a Subscription, no Notification can
slip in between.
- This sort of is a replacement for the below call
synchronizeSubscriberByFCP(): The old implementation of that did both
the job of creating the synchronization and sending it out to the
client. That call is not removed yet though because it won't compile
like this anymore anyway, so people will notice it, and it is more
suitable to remove it in a following commit which will deal with it in
the right place.