SubscriptionManager.Subscription: Be generic of type extending EventSource
(EventSource = interface which Identity / Trust / Score implement.)

- This is better than the previous type parameter of extending
Notification: It will allow storage of Notification objects of different
types for a given subscription. This is needed to refactor the code of
subscription synchronizations to store them as objects of a special
Notification class - which will make the code a lot cleaner.
- I've made sure that the JavaDoc of all modified code is corrected.
- I've reviewed all changed functions to make sure that their callers /
callees are safe with the new type parameter.
- Line lengths are also fixed as I was touching the code anyway.